refactoring extract errors to consts & rename

+++ Api/Controllers/UserTreeNodeController.cs
@@ -12,12 +12,10 @@ namespace Data.Controllers
     [Route("api.user.tree.node")]
     public class UserTreeNodeController : ControllerBase
     {
-        private readonly ILogger _logger;
         private readonly ApplicationDbContext _dbContext;
 
-        public UserTreeNodeController(ILoggerProvider loggerProvider, ApplicationDbContext dbContext)
+        public UserTreeNodeController(ApplicationDbContext dbContext)
         {
-            _logger = loggerProvider.CreateLogger("JournalMessageController") ?? throw new ArgumentNullException(nameof(loggerProvider));
             _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
         }
         /// <remarks>
@@ -25,14 +23,14 @@ namespace Data.Controllers
         /// </remarks>
         [Route("/api.user.tree.node.create")]
         [HttpPost]
-        public async Task<IActionResult> CreateAsync([FromQuery] TreeNodeCreateQueryStringParameters args)
+        public async Task<IActionResult> CreateAsync([FromQuery] UserTreeNodeCreateQueryStringParameters args)
         {
             var tree = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.Id == args.parentNodeId && x.TreeName == args.treeName);
             if (tree == null)
-                throw new SecureException("Parent node dosn't exist");
+                throw new SecureException(UserTreeNodeErrors.ParentNotFound);
             var node = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.ParentNodeId == args.parentNodeId && x.TreeName == args.treeName && x.Name == args.nodeName);
             if (node != null)
-                throw new SecureException("Node alredy exist");
+                throw new SecureException(UserTreeNodeErrors.NodeAlredyExist);
 
             node = new tree_node()
             {
@@ -50,15 +48,18 @@ namespace Data.Controllers
         /// </remarks>
         [Route("/api.user.tree.node.delete")]
         [HttpPost]
-        public async Task<IActionResult> DeleteAsync([FromQuery] TreeNodeDeleteQueryStringParameters args)
+        public async Task<IActionResult> DeleteAsync([FromQuery] UserTreeNodeDeleteQueryStringParameters args)
         {
+            var rootNode = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.ParentNodeId == null && x.TreeName == args.treeName);
+            if (rootNode == null)
+                throw new SecureException(UserTreeNodeErrors.TreeNotFound);
             var anyNodeChild = _dbContext.TreeNodes.AsNoTracking().Any(x => x.ParentNodeId == args.nodeId && x.TreeName == args.treeName);
             if (anyNodeChild)
-                throw new SecureException("You have to delete all children nodes first");
+                throw new SecureException(UserTreeNodeErrors.ChildNodesShouldBeDeletedFirst);
 
             var node = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.Id == args.nodeId && x.TreeName == args.treeName);
             if (node != null)
-                throw new SecureException("Tree dosn't exist");
+                throw new SecureException(UserTreeNodeErrors.TreeNodeNotFound);
 
             _dbContext.TreeNodes.Remove(node);
             await _dbContext.SaveChangesAsync();
@@ -70,11 +71,11 @@ namespace Data.Controllers
         /// </remarks>
         [Route("/api.user.tree.node.rename")]
         [HttpPost]
-        public async Task<IActionResult> RenameAsync([FromQuery] TreeNodeRenameQueryStringParameters args)
+        public async Task<IActionResult> RenameAsync([FromQuery] UserTreeNodeRenameQueryStringParameters args)
         {
-            var node = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.Id == args.nodeId && x.Name == args.treeName);
+            var node = await _dbContext.TreeNodes.AsNoTracking().FirstOrDefaultAsync(x => x.Id == args.nodeId && x.TreeName == args.treeName);
             if (node != null)
-                throw new SecureException("Tree dosn't exist");
+                throw new SecureException(UserTreeNodeErrors.TreeNodeNotFound);
             node.Name = args.newNodeName;
             _dbContext.TreeNodes.Update(node);
             await _dbContext.SaveChangesAsync();
